Queue: Handler and Manager separation refactor
