+++ qcfractal/queue_handlers/__init__.py
@@ -1,1 +1,6 @@
-from .queue_handlers import build_queue
+"""
+Initializer for the queue_handler folder
+"""
+
+from .queue_managers import build_queue_manager
+from .queue_handlers import QueueScheduler, ServiceScheduler
